Unity build: single-TU compilation for full optimizer visibility
All decoder code now compiles as one translation unit via
sub0h264_unity.cpp. This gives the compiler complete visibility
across all modules (NAL, SPS/PPS, CAVLC, transform, prediction,
deblocking, DPB) for maximum inlining and optimization — critical
for ESP32-P4 where LTO is unavailable.

- sub0h264_unity.cpp: Single TU that includes decoder.hpp (which
  transitively includes all decoder headers) + public API stubs
- CMakeLists.txt: Updated to build from unity source, added PRIVATE
  include path for src/ directory
- Fixed std::fill uint8_t/unsigned int conversion warning

90 test cases, all passing.

+++ components/sub0h264/src/decoder.hpp
@@ -261,9 +261,9 @@ private:
             decodeTarget = dpb_.getDecodeTarget();
         }
 
-        std::fill(nnzLuma_.begin(), nnzLuma_.end(), 0U);
-        std::fill(nnzCb_.begin(), nnzCb_.end(), 0U);
-        std::fill(nnzCr_.begin(), nnzCr_.end(), 0U);
+        std::fill(nnzLuma_.begin(), nnzLuma_.end(), static_cast<uint8_t>(0U));
+        std::fill(nnzCb_.begin(), nnzCb_.end(), static_cast<uint8_t>(0U));
+        std::fill(nnzCr_.begin(), nnzCr_.end(), static_cast<uint8_t>(0U));
         std::fill(mbMotion_.begin(), mbMotion_.end(), MbMotionInfo{});
 
         // Compute effective QP
